[#30] refactor: 허브 관련 에러 패키지 생성 및 위치 변경
