+++ com.jv-cc.flowmesh.company/src/main/java/com/jv_cc/flowmesh/company/application/exception/hub/HubException.java
@@ -1,12 +1,13 @@
-package com.jv_cc.flowmesh.company.application.exception;
+package com.jv_cc.flowmesh.company.application.exception.hub;
 
+import com.jv_cc.flowmesh.company.application.exception.Error;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 @Getter
 public class HubException extends RuntimeException {
 
-    Error error;
+    com.jv_cc.flowmesh.company.application.exception.Error error;
 
     HttpStatus httpStatus;
