+++ com.jv-cc.flowmesh.company/src/main/java/com/jv_cc/flowmesh/company/application/exception/hub/NotFoundHubException.java
@@ -1,5 +1,6 @@
-package com.jv_cc.flowmesh.company.application.exception;
+package com.jv_cc.flowmesh.company.application.exception.hub;
 
+import com.jv_cc.flowmesh.company.application.exception.Error;
 import org.springframework.http.HttpStatus;
 
 public class NotFoundHubException extends HubException {
